Demonstrating Trace hooking
Trace system seems to be a very extensive debugging framework of some sort, but I haven't yet spent much time to look into it. It looks like it can be very verbose, but it isn't obvious how to get the data out.
For testing purposes, I've added a simple LD_PRELOAD hooking demo that redirects tracing messages to stdout instead of /dev/trace.